【modify】conversation.pyを作成し、LangGraphを用いたチャットのやり取りを別ファイルで管理する
