+++ web_search_agent.py
@@ -1,5 +1,5 @@
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
-from langchain_core.messages import HumanMessage, SystemMessage
+from langchain_core.messages import SystemMessage
 
 from langchain_google_genai import ChatGoogleGenerativeAI
 
@@ -9,6 +9,7 @@ from langgraph.checkpoint.memory import MemorySaver
 from langgraph.graph import START, END, StateGraph
 from langgraph.prebuilt import ToolNode
 
+from conversation import conversation
 from prompts import web_search_system_prompt
 from state import GraphState
 
@@ -43,35 +44,6 @@ def create_langgraph(chain, tools):
     memory = MemorySaver()
     graph = workflow.compile(checkpointer=memory)
     return graph
-
-def conversation(graph):
-    while True:
-        query = input("質問を入力してください: ")
-
-        if query.lower() in ["exit", "quit"]:
-            print("終了します。")
-            break
-
-        print("=================================")
-        print("質問:", query)
-
-        input_query = [HumanMessage(
-            [
-                {
-                    "type": "text",
-                    "text": f"{query}"
-                },
-            ]
-        )]
-
-        # 同じスレッドIDでinvokeが繰り返されることで、会話履歴が引き継がれる
-        response = graph.invoke({"messages": input_query}, config={"configurable": {"thread_id": "12345"}})
-
-        # デバック用
-        print("response: ", response)
-
-        print("=================================")
-        print("AIの回答", response["messages"][-1].content)
 
 def main():
     # orieg/gemma3-tools:27b-it-qatだと性能不足でtoolを活用することが出来なかった
